requirements.txt
# Data Processing & Analysis
pandas==2.2.0
numpy==1.26.0

# Database Connectivity
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Synthetic Data Generation
faker==22.6.0

# Testing & Environment
pytest==8.0.0
python-dotenv==1.0.1
pyyaml==6.0.1